refactor(core): split cart item update use case
- move shared cart loading, normalization, and snapshot helpers into cartState
- extract updateCartItem into its own use-case module while keeping barrel exports stable
- reduce cart.ts hotspot score from 34.1 to 24.3 and density from 0.14 to 0.10
