+++ packages/coffee/core/src/application/use-cases/index.ts
@@ -3,7 +3,7 @@
  *
  * @module
  */
-export { addCartItem, clearCart, getCart, removeCartItem, updateCartItem } from "./cart.ts";
+export { addCartItem, clearCart, getCart, removeCartItem } from "./cart.ts";
 export { checkoutCart } from "./checkoutCart.ts";
 export { markReady, cancelOrder, pickUpOrder, startBrewing } from "./changeOrderStatus.ts";
 export { getCurrentCheckoutSession, prepareCartCheckout } from "./checkoutSession.ts";
@@ -13,3 +13,4 @@ export { listMenu } from "./listMenu.ts";
 export { listOrders } from "./listOrders.ts";
 export { placeOrder } from "./placeOrder.ts";
 export { quoteOrder, validateOrder } from "./quoteOrder.ts";
+export { updateCartItem } from "./updateCartItem.ts";
